correctly matched column.type in column definition and column result variable type for "delta_MR" field
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/goby@30413 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/goby/modes/MethylationRateVCFOutputFormat.java
@@ -330,7 +330,7 @@ public class MethylationRateVCFOutputFormat implements SequenceVariationOutputFo
                 group2MR=methylatedCCountPerGroup[indexGroup2]*100/totalCsObservedgroup2;
             }
 
-            final double deltaMR=Math.round(Math.abs(group1MR-group2MR));
+            final int deltaMR= (int) Math.round(Math.abs(group1MR-group2MR));
 
             statWriter.setInfo(log2OddsRatioColumnIndex[comparison.index], log2OddsRatio);
             statWriter.setInfo(log2OddsRatioStandardErrorColumnIndex[comparison.index], logOddsRatioSE);
